requirements.txt
django
djangorestframework
factory-boy
django-rest-assured